NoxOS/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
sqlalchemy==2.0.34
pymysql==1.1.0
cryptography==42.0.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.12.5
pydantic-settings==2.1.0
python-dotenv==1.0.1
